fixed routing and added error messages and other minor fixes
