+++ src/app/(private)/completeProfile/wazaWarrior/[user_id]/page.tsx
@@ -1,42 +1,31 @@
 'use client';
-import { useState, useEffect } from 'react';
-import { useSession } from 'next-auth/react';
-import { useRouter } from 'next/navigation';
+import { signIn } from 'next-auth/react';
+import { useParams, useRouter } from 'next/navigation';
+import { useEffect, useState } from 'react';
 import userIcon from '@/assets/formIcons/user.svg';
 import wazaLogoBlack from '@/assets/wazaLogos/Wazalogo_Black.svg';
 import Image from 'next/image';
 
-export default function CompleteUserPage() {
+export default function CompleteWarriorProfile() {
   const router = useRouter();
-  const { data: sessionData, status } = useSession();
-  const [userDetails, setUserDetails] = useState({
-    username: '',
-    user_type: '',
-    email: sessionData?.user.email || '',
-    profile_pic: sessionData?.user.image || '',
-    is_verified: sessionData?.user.is_verified || false,
-    provider: sessionData?.user.provider || '',
+  const { user_id } = useParams<{ user_id: string }>() || { user_id: '' };
+  const [warriorDetails, setWarriorDetails] = useState({
+    user_id: user_id,
   });
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [error, setError] = useState('');
 
-  // Update userDetails with session data when it's loaded
   useEffect(() => {
-    if (status === 'authenticated') {
-      setUserDetails((prevDetails) => ({
-        ...prevDetails,
-        email: sessionData.user.email || '',
-        profile_pic: sessionData.user.image || '',
-        is_verified: sessionData.user.is_verified || false,
-        provider: sessionData.user.provider || '',
-      }));
-    }
-  }, [sessionData, status]);
+    setWarriorDetails((prevDetails) => ({
+      ...prevDetails,
+      user_id: user_id,
+    }));
+  }, [user_id]);
 
   useEffect(() => {
     (async () => {
       const res = await fetch(
-        `http://localhost:3000/api/user?email=${sessionData?.user.email}`,
+        `http://localhost:3000/api/waza_warrior?user_id=${user_id}`,
         {
           method: 'GET',
           headers: {
@@ -46,24 +35,26 @@ export default function CompleteUserPage() {
       );
       if (res.ok) {
         const data = await res.json();
-        console.log('User found:', data);
-        if (data.user_type === 'Waza Trainer') {
-          router.push(`/complete-user/waza_trainer/${data.user_id}`);
-        } else if (data.user_type === 'Waza Warrior') {
-          router.push(`/complete-user/waza_warrior/${data.user_id}`);
-        }
+        console.log('Warrior found:', data);
+        router.push(`/dashboard`);
       }
     })();
-  }, [sessionData, router]);
+  }, [user_id, router]);
 
-  // Handle input changes for each field
-  const handleInputChange = (e: any) => {
-    const { id, value, type, checked } = e.target;
-    setUserDetails((prevDetails) => ({
-      ...prevDetails,
-      [id]: type === 'checkbox' ? checked : value,
-    }));
-  };
+  useEffect(() => {
+    (async () => {
+      const res = await fetch(`http://localhost:3000/api/user?id=${user_id}`, {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      });
+      if (!res.ok) {
+        console.log('User Not found:');
+        router.push(`/completeProfile`);
+      }
+    })();
+  }, [user_id, router]);
 
   const handleSubmit = async (e: any) => {
     e.preventDefault();
@@ -71,32 +62,31 @@ export default function CompleteUserPage() {
     setError('');
 
     try {
-      console.log('User details:', userDetails);
-      const response = await fetch(`http://localhost:3000/api/user`, {
+      const response = await fetch(`http://localhost:3000/api/waza_warrior`, {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
-        body: JSON.stringify(userDetails),
+        body: JSON.stringify(warriorDetails),
       });
 
       const data = await response.json();
-      console.log(data);
       if (!response.ok) {
         throw new Error(data.message || 'Something went wrong!');
       }
-      console.log('User created:', data);
-      if (data.user_type === 'Waza Warrior')
-        router.push(`complete-user/waza_warrior/${data.user_id}`);
-      else if (data.user_type === 'Waza Trainer')
-        router.push(`complete-user/waza_trainer/${data.user_id}`);
-    } catch (err: any) {
-      console.error('Error creating user:', err);
-      setError(err.message);
+      console.log('Warrior created:', data);
+      signIn(); // Use router.navigate for Next.js 13+
+    } catch (err) {
+      // If err is an instance of Error, use its message, otherwise use a default error message
+      const errorMessage =
+        err instanceof Error ? err.message : 'An unknown error occurred';
+      console.error('Error creating warrior:', errorMessage);
+      setError(errorMessage);
     } finally {
       setIsSubmitting(false);
     }
   };
+
   return (
     <>
       <div className='flex flex-col items-center justify-center min-h-screen'>
@@ -111,9 +101,23 @@ export default function CompleteUserPage() {
             </div>
             <input
               type='text'
-              id='username'
-              name='username'
-              onChange={handleInputChange}
+              id='name'
+              name='name'
+              // onChange={handleInputChange}
+              className='bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg block focus:ring-yellow-400 focus:border-yellow-400 w-full p-2 ps-10'
+              placeholder='Username'
+              required
+            />
+          </div>
+          <div className='relative mb-5'>
+            <div className='absolute inset-y-0 start-0 flex items-center ps-2.5 pointer-events-none'>
+              <Image src={userIcon} alt='userIcon' />
+            </div>
+            <input
+              type='number'
+              id='age'
+              name='age'
+              // onChange={handleInputChange}
               className='bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg block focus:ring-yellow-400 focus:border-yellow-400 w-full p-2 ps-10'
               placeholder='Username'
               required
